test: delete question

+++ src/domain/forum/application/use-cases/delete-question.ts
@@ -8,12 +8,17 @@ export class DeleteQuestionUseCase {
   constructor(private questionRepository: IQuestionsRepository) {}
 
   async execute({
+    authorId,
     questionId,
   }: IDeleteQuestionUseCaseRequest): Promise<IDeleteQuestionUseCaseResponse> {
     const question = await this.questionRepository.findById(questionId)
 
     if (!question) {
       throw new Error('Question not found.')
+    }
+
+    if (authorId !== question.authorId.toString()) {
+      throw new Error('Not allowed.')
     }
 
     await this.questionRepository.delete(question)
